requirements.proxy.txt
psutil==5.9.5
python-json-logger==2.0.7
prometheus_client>=0.16.0
pyzmq>=25.0.0